app/courses/serializers.py
from rest_framework import serializers
from courses.models import Course


class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = ('id', 'title', 'course_url', 'image_path', 'description',
                  'published')
