app/requirements.txt
django==3.2.2
gunicorn==20.0.4
djangorestframework==3.12.4
pytest==6.2.5
pytest-django==4.4.0
python-decouple==3.4
psycopg2-binary==2.9.6